extract breadcrumb logic to layout
